fix: api error with harmful video
