+++ src/scripts/api/nicoad.ts
@@ -20,11 +20,13 @@ export interface FetchNicoadResponse {
 }
 
 // 公式動画の場合、watchIdは'so***'に変換したもの
-export function fetchNicoad(watchId: string) {
+export async function fetchNicoad(
+  watchId: string
+): Promise<FetchNicoadResponse> {
   const uri = `https://api.nicoad.nicovideo.jp/v1/contents/video/${watchId}`;
   const options = {
     method: "GET",
   };
 
-  return fetch(uri, options);
+  return fetch(uri, options).then((response) => response.json());
 }
